Rename observe module to api
