+++ switchy/__init__.py
@@ -9,7 +9,7 @@ Licensed under the MPL 2.0 license (see `LICENSE` file)
 from os import path
 from . import apps
 from .utils import get_logger, ESLError
-from .observe import Client
+from .api import Client
 from .async import get_event_loop
 from .handlers import get_listener
 from .apps.call_gen import get_originator
